requirements.txt
fastapi==0.115.6
python-multipart==0.0.17
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.6.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.14.0
pyotp==2.9.0
numpy>=2.0.0
opencv-python-headless>=4.10.0
pillow==10.3.0
bcrypt==4.0.1
reportlab>=4.0
openpyxl>=3.1.0
python-docx>=1.0.0
